feat(bsppp): fuse `vpkpp::BSP` and `bsppp::BSP`
